backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow>=10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
